Pull request to main

+++ server/controllers/uploads.js
@@ -1,5 +1,8 @@
 const imageController = require("./image");
 const Uploads = require("../models/uploads")
+const bcrypt = require('bcrypt');
+const saltRounds = 10;
+const fs = require('fs');
 
 exports.getUploads = async (req, res) => {
   try {
@@ -32,30 +35,66 @@ exports.getUpload = async (req, res) => {
       }
 };
 
+exports.searchForUpload = async (req, res) => {
+     try {
+        const {search} = req.body
+        console.log(search)
+        const {category} = req.params
+        if (!search || !category) return res.status(400).send({msg: 'Something is missing'})
+        const result = await Uploads.find({name: new RegExp(search, "i"), category: category});
+        if(result){
+            return res.status(200).send({
+                msg: "Upload found!",
+                payload: result,
+            })
+        }
+        res.status(500).send({msg: "Something went wrong"});
+      } catch (error) {
+        console.log(error);
+        res.status(500).send(error);
+      }
+}
+
 const uploadFile = imageController.upload.single("imageFile");
 
 const saveFileIntoFolder = (req, res, next) => {
-    uploadFile(req, res, (err) => {
+     uploadFile(req, res, (err) => {
       if(err){
-        console.log(err);
+            switch(err.message) {
+                case "Something is missing":
+                    return res.status(400).send({msg: 'Something is missing'})
+                case "password should be the same":
+                    return res.status(400).send({msg: 'Passwords should be same'})
+                case "contact should be type number":
+                    return res.status(400).send({msg: 'Contact should be type number'})
+                case "price should be type number":
+                    return res.status(400).send({msg: 'Price should be type number'})
+                case "File not found":
+                    return res.status(400).send({msg: 'File not found'})
+            }
             return res.status(500).send(err);
       }
-      console.log("File uploaded!")
+        console.log("File uploaded!")
         next();
     });
 }
 
 const saveIntoDb = async (req, res) => {
+ 
     try {
+        if (!req.file) return res.status(400).send({msg: 'File not found'})
         const upload = new Uploads({
             name: req.body.name,
             contact: req.body.contact,
             location: req.body.location,
             nameOfSeller: req.body.nameOfSeller,
             price: req.body.price,
-            //imagePath: "http://localhost:3000/img/" + req.file.filename,
+            description: req.body.description,
+            category: req.body.category,
+            password: bcrypt.hashSync(req.body.password, saltRounds),
+            imagePath: "http://localhost:3000/img/" + req.file.filename,
         });
-        const result = await upload.save();
+        const result = await upload.save(); 
         if(result){
             return res.status(201).send({
                 msg: "Upload uploaded",
@@ -63,7 +102,6 @@ const saveIntoDb = async (req, res) => {
             })
         }
         return res.status(500).send({msg: "nejde nic"});
-
     } catch (error) {
         console.log(error);
         res.status(500).send({
@@ -80,6 +118,9 @@ exports.updateUpload = async (req, res) => {
             location: req.body.location,
             nameOfSeller: req.body.nameOfSeller,
             price: req.body.price,
+            description: req.body.description,
+            category: req.body.category,
+            password: bcrypt.hashSync(req.body.password, saltRounds),
             //imagePath: "http://localhost:3000/img/" + req.file.filename,
         };
         const result = await Uploads.findByIdAndUpdate(req.params.id, upload);
@@ -98,9 +139,37 @@ exports.updateUpload = async (req, res) => {
     }
 }
 
-exports.deleteUpload = async (req, res) => {
+const deleteUpload = async (req, res, next) => {
     try {
-        const data = await Uploads.findByIdAndDelete(req.params.id);
+        console.log(req.body)
+        if(!req.body.password) return res.status(400).send({msg: "Something is missing"});
+        
+        const data = await Uploads.findById(req.params.id);
+        const match = await bcrypt.compare(req.body.password, data.password);
+
+        if(!match) return res.status(400).send({msg: "Passwords do not match"});
+        const deletedUpload = await Uploads.findByIdAndDelete(req.params.id);
+
+        const name = deletedUpload.imagePath.split('/');
+        console.log(name[4])
+        fs.unlink(__dirname + `/../public/img/${name[4]}`, (err) => {
+            if(err) throw err;
+            console.log("Image deleted")
+        })
+
+        if(!deletedUpload) return res.status(500).send({msg: "Something went wrong"});
+        console.log(deletedUpload)
+        return res.status(200).send({msg: "Succesful", payload: deletedUpload});
+    } catch (error) {
+        res.status(500).send({
+            error,
+        });
+    }
+}
+
+exports.deleteAllUploads = async (req, res) => {
+    try {
+        const data = await Uploads.deleteMany();
         if(data){
             return res.status(200).send({
                 msg: "Upload deleted",
@@ -117,5 +186,6 @@ exports.deleteUpload = async (req, res) => {
     }
 }
 
-
 exports.postUpload = [saveFileIntoFolder, saveIntoDb];
+exports.deleteUpload = [deleteUpload]
+
